pyproject.toml
[project]
name = "ft-cookbook"
version = "0.1.0"
description = "Recipes for building web apps with FastHTML"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ft-jupyter",
    "python-fasthtml>=0.10.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "nbdev>=2.3.34",
    "uv>=0.5.7",
]

[tool.uv.sources]
ft-jupyter = { git = "https://github.com/banditburai/ft-jupyter.git" }
